maintain: remove timestamp and caller from console logs
Also move the file logger into logger.go so that all functions that modify the static
L are in the same place.

Also adds godoc for all the functions that modify L.
